LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-track-settings.tentative-expected.txt


FAIL RTCRtpReceiver.track.getSettings() contains resolution entries assert_equals: Receive track width setting is updated after resolution change expected 240 but got 640
